apps/dashboard/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
requests==2.32.3
kafka-python==2.0.2
